revert: undo all AIW-3 changes (poll workflow migration)
Reverts the cron-to-workflow migration made in AIW-3, restoring
the original cron-based polling mechanism.
