+++ .env.example
@@ -38,8 +38,8 @@ POLL_INTERVAL_MS=300000
 # VERCEL_TEAM_ID=
 # VERCEL_PROJECT_ID=
 
-# Deploy hook auth
-DEPLOY_HOOK_SECRET=
+# Cron auth
+CRON_SECRET=
 
 # Workflow (local dev only)
 WORKFLOW_POSTGRES_URL=postgresql://localhost:5432/ai_workflow
