+++ env.ts
@@ -36,15 +36,15 @@ export const env = createEnv({
     JOB_TIMEOUT_MS: z.coerce.number().int().positive().default(1_800_000),
 
     // Polling
-    POLL_INTERVAL_MS: z.coerce.number().int().positive().default(15_000),
-
-    // Deploy hook auth
-    DEPLOY_HOOK_SECRET: z.string().min(1).optional(),
+    POLL_INTERVAL_MS: z.coerce.number().int().positive().default(300_000),
 
     // Vercel (optional — auto via OIDC on Vercel)
     VERCEL_TOKEN: z.string().min(1).optional(),
     VERCEL_TEAM_ID: z.string().min(1).optional(),
     VERCEL_PROJECT_ID: z.string().min(1).optional(),
+
+    // Cron
+    CRON_SECRET: z.string().min(1).optional(),
 
     // Redis (run registry)
     AI_WORKFLOW_KV_REST_API_URL: z.string().url(),
