+++ vercel.json
@@ -1,1 +1,8 @@
-{}
+{
+  "crons": [
+    {
+      "path": "/cron/poll",
+      "schedule": "* * * * *"
+    }
+  ]
+}
